(feat) Document and complete MVP features

+++ src/query.ts
@@ -1,44 +1,108 @@
-import type { Model, ModelFieldKey, ModelFields } from './models'
+import type { Filter, ModelFieldKey, OrderBy } from './types'
+import type { Model } from './models'
 import { _objectStore } from './connection'
 
-export type Filter<T extends Model> = Partial<ModelFields<T>>
-export type OrderBy<T extends Model> = (ModelFieldKey<T>) | `-${ModelFieldKey<T>}`
-
 export class Query<T extends Model> {
+  /** The filters to apply to the query */
   private filters = {} as Filter<T>
+  /** The field to order the query by */
   private _orderBy?: ModelFieldKey<T>
-  private reverse = false
+  /** Whether to reverse the order of the query */
+  private _reverse = false
 
-  constructor(private TargetModel: { new(): T }) {}
+  /**
+   * A query builder for the models.
+   *
+   * A query is lazy, meaning that it will not execute until you call one of the methods.
+   *
+   * A query is linked to a model, meaning that you can only query the model that you created the query from.
+   *
+   * @see {@link Model}: {@link Model.filter} and {@link Model.orderBy} for how queries are usually created
+   *
+   * @typeParam T - The model to query
+   * @param TargetModel - The model to query
+   */
+  public constructor(private TargetModel: { new(): T }) {}
 
-  filter(filters: Filter<T>): Query<T> {
+  /**
+   * Update the filters of the query.
+   * This will not reset the filters, but will merge the new filters with the old ones.
+   * @param filters - The new filters to apply to the query
+   * @returns - The query itself, to allow chaining
+   */
+  public filter(filters: Filter<T>): Query<T> {
     Object.assign(this.filters, filters)
     return this
   }
 
-  orderBy(field: OrderBy<T>): Query<T> {
+  /**
+   * Reset the filters of the query.
+   * @returns - The query itself, to allow chaining
+   */
+  public resetFilters(): Query<T> {
+    this.filters = {} as Filter<T>
+    return this
+  }
+
+  /**
+   * Update the field used to order the query.
+   * @param field - The field to order the query by, prepend with `-` to reverse the order
+   * @returns - The query itself, to allow chaining
+   */
+  public orderBy(field: OrderBy<T>): Query<T> {
     if (field.startsWith('-')) {
-      this.reverse = true
+      this._reverse = true
       this._orderBy = field.slice(1) as ModelFieldKey<T>
     }
     else {
+      this._reverse = false
       this._orderBy = field as ModelFieldKey<T>
     }
     return this
   }
 
-  _getCursor() {
+  /**
+   * Change the order of the query.
+   * @returns - The query itself, to allow chaining
+   */
+  public reverse(): Query<T> {
+    this._reverse = !this._reverse
+    return this
+  }
+
+  /**
+   * Checks if the given instance fits the filters of the query.
+   * @param instance - The instance to check
+   * @returns - Whether the instance fits the filters
+   */
+  private _fitsFilters(instance: T): boolean {
+    for (const [key, value] of Object.entries(this.filters)) {
+      if (instance[key as keyof T] !== value as T[keyof T])
+        return false
+    }
+    return true
+  }
+
+  /**
+   * Utility function to get the cursor of the query.
+   * @returns
+   */
+  private _getCursor(): IDBRequest<IDBCursorWithValue | null> {
     const store = _objectStore(this.TargetModel.name)
     if (this._orderBy) {
       const index = store.index(this._orderBy)
 
-      return index.openCursor(undefined, this.reverse ? 'prev' : 'next')
+      return index.openCursor(undefined, this._reverse ? 'prev' : 'next')
     }
     else {
-      return store.openCursor()
+      return store.openCursor(undefined, this._reverse ? 'prev' : 'next')
     }
   }
 
+  /**
+   * Executes the query and returns the first result.
+   * @returns - The first result of the query, or null if no result was found
+   */
   async first(): Promise<T | null> {
     const cursor = this._getCursor()
     return new Promise<T | null>((resolve, reject) => {
@@ -47,9 +111,12 @@ export class Query<T extends Model> {
           resolve(null)
         }
         else {
-          const instance = new this.TargetModel()
-          Object.assign(instance, cursor.result.value)
-          resolve(instance)
+          if (this._fitsFilters(cursor.result.value as T)) {
+            const instance = new this.TargetModel()
+            Object.assign(instance, cursor.result.value)
+            resolve(instance)
+          }
+          else { cursor.result.continue() }
         }
       }
       cursor.onerror = (event) => {
@@ -58,6 +125,10 @@ export class Query<T extends Model> {
     })
   }
 
+  /**
+   * Executes the query and returns all the results.
+   * @returns - All the results of the query
+   */
   async all(): Promise<T[]> {
     const cursor = this._getCursor()
     const result: T[] = []
@@ -67,9 +138,11 @@ export class Query<T extends Model> {
           resolve(result)
         }
         else {
-          const instance = new this.TargetModel()
-          Object.assign(instance, cursor.result.value)
-          result.push(instance)
+          if (this._fitsFilters(cursor.result.value as T)) {
+            const instance = new this.TargetModel()
+            Object.assign(instance, cursor.result.value)
+            result.push(instance)
+          }
           cursor.result.continue()
         }
       }
@@ -79,6 +152,10 @@ export class Query<T extends Model> {
     })
   }
 
+  /**
+   * Executes the query and returns the number of results.
+   * @returns - The amount of results of the query
+   */
   async count(): Promise<number> {
     const cursor = this._getCursor()
     let count = 0
@@ -88,7 +165,9 @@ export class Query<T extends Model> {
           resolve(count)
         }
         else {
-          count++
+          if (this._fitsFilters(cursor.result.value as T))
+            count++
+
           cursor.result.continue()
         }
       }
@@ -98,15 +177,23 @@ export class Query<T extends Model> {
     })
   }
 
-  async delete(): Promise<void> {
+  /**
+   * Executes the query and deletes all the results.
+   * @returns - The amount of results deleted
+   */
+  async delete(): Promise<number> {
     const cursor = this._getCursor()
-    return new Promise<void>((resolve, reject) => {
+    let amount = 0
+    return new Promise<number>((resolve, reject) => {
       cursor.onsuccess = () => {
         if (!cursor.result) {
-          resolve()
+          resolve(amount)
         }
         else {
-          cursor.result.delete()
+          if (this._fitsFilters(cursor.result.value as T)) {
+            amount += 1
+            cursor.result.delete()
+          }
           cursor.result.continue()
         }
       }
